use unified API for GraalJS and RhinoJS for parsing response headers
