+++ maestro-test/src/test/kotlin/maestro/test/JsEngineTest.kt
@@ -1,5 +1,6 @@
 package maestro.test
 
+import com.github.tomakehurst.wiremock.client.WireMock
 import com.github.tomakehurst.wiremock.client.WireMock.equalTo
 import com.github.tomakehurst.wiremock.client.WireMock.equalToJson
 import com.github.tomakehurst.wiremock.client.WireMock.get
@@ -8,6 +9,7 @@ import com.github.tomakehurst.wiremock.client.WireMock.post
 import com.github.tomakehurst.wiremock.client.WireMock.stubFor
 import com.github.tomakehurst.wiremock.junit5.WireMockRuntimeInfo
 import com.github.tomakehurst.wiremock.junit5.WireMockTest
+import com.google.common.net.HttpHeaders
 import com.google.common.truth.Truth.assertThat
 import maestro.js.JsEngine
 import org.junit.jupiter.api.Test
@@ -164,5 +166,46 @@ abstract class JsEngineTest {
 
         result = engine.evaluateScript("FOO").toString()
         assertThat(result).isEqualTo("foo")
+    }
+
+    @Test
+    fun `HTTP - Make GET request and check response body and headers `(wiremockInfo: WireMockRuntimeInfo) {
+        // Given
+        val port = wiremockInfo.httpPort
+        val body =
+            """
+                {
+                    "message": "GET Endpoint"
+                }
+            """.trimIndent()
+
+        val testHeader = "testHeader"
+        val response = WireMock.aResponse().withStatus(200)
+            .withHeader(HttpHeaders.CONTENT_TYPE, "application/json")
+            .withHeader(testHeader, "first")
+            .withHeader(testHeader, "second")
+            .withBody(body);
+
+        WireMock.stubFor(
+            WireMock.get("/json").willReturn(response)
+        )
+
+        val script = """
+            var response = http.get('http://localhost:$port/json');
+            
+            //check body
+            var message = json(response.body).message;
+            
+            // check headers
+            var contentType = response.headers['content-type'];
+            var testHeader = response.headers['testheader'];
+            String(message + String(" ") + contentType + String(" ") + testHeader);
+        """.trimIndent()
+
+        // When
+        val result = engine.evaluateScript(script)
+
+        // Then
+        assertThat(result.toString()).isEqualTo("GET Endpoint application/json first,second")
     }
 }
