+++ maestro-test/src/test/kotlin/maestro/test/RhinoJsEngineTest.kt
@@ -58,46 +58,4 @@ class RhinoJsEngineTest : JsEngineTest() {
         val result = engine.evaluateScript("parseInt('1')").toString()
         assertThat(result).isEqualTo("1.0")
     }
-
-    @Test
-    fun `HTTP - Make GET request and check response body and headers `(wiremockInfo: WireMockRuntimeInfo) {
-        // Given
-        val port = wiremockInfo.httpPort
-        val body =
-            """
-                {
-                    "message": "GET Endpoint"
-                }
-            """.trimIndent()
-
-        val testHeader = "testHeader"
-        val response = WireMock.aResponse().withStatus(200)
-            .withHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-            .withHeader(testHeader, "first")
-            .withHeader(testHeader, "second")
-            .withBody(body);
-
-        WireMock.stubFor(
-            WireMock.get("/json").willReturn(response)
-        )
-
-        val script = """
-            var response = http.get('http://localhost:$port/json');
-            
-            //check body
-            var message = json(response.body).message;
-            
-            
-            // check headers
-            var contentType = response.headers.get("content-type");
-            var testHeader = response.headers.get("testHeader");
-            String(message + String(" ") + contentType.get(0)) + String(" ") + String(testHeader.get(0)) + String(" ") + String(testHeader.get(1));
-        """.trimIndent()
-
-        // When
-        val result = engine.evaluateScript(script)
-
-        // Then
-        assertThat(result.toString()).isEqualTo("GET Endpoint application/json first second")
-    }
 }
